update export ruby to not fail on missing index.html , just log error

+++ _docker/lib/export/export_html_post_processor.rb
@@ -207,10 +207,16 @@ class ExportHtmlPostProcessor
   # to the root of the directory
   #
   def relocate_index_html(export_directory)
-    FileUtils.mv("#{export_directory}/index/index.html", "#{export_directory}/index.html")
-    FileUtils.rm_rf("#{export_directory}/index")
-    @process_runner.execute!("sed -i'' -e s:'\\.\\.\\/':'':g #{export_directory}/index.html")
-    @log.info("Moved #{export_directory}/index/index.html to #{export_directory}/index.html.")
+    
+    if File.exists?("#{export_directory}/index/index.html") 
+      FileUtils.mv("#{export_directory}/index/index.html", "#{export_directory}/index.html")
+      FileUtils.rm_rf("#{export_directory}/index")
+      @process_runner.execute!("sed -i'' -e s:'\\.\\.\\/':'':g #{export_directory}/index.html")
+      @log.info("Moved #{export_directory}/index/index.html to #{export_directory}/index.html.")
+    else 
+      @log.warn("Failed to move #{export_directory}/index/index.html to #{export_directory}/index.html. source file missing")
+    end
+
   end
 
   #
